include/nutconf.hpp: align UpsmonConfiguration::notifyFlags type with clients/upsmon.h (notifylist->flags)
Note that here we have "unsigned int" and there a plain "int" at the moment;
but even with short signed ints we currently do not have too many values :)

This fix also drops a warning about shortening the cast when assigning from
a temporary "unsigned int flags" in UpsmonConfigParser::onParseDirective()

+++ common/nutwriter.cpp
@@ -253,7 +253,7 @@ const NotifyFlagsStrings::FlagStrings NotifyFlagsStrings::flag_str =
  *
  *  \return NOTIFYFLAG directive string
  */
-static std::string serializeNotifyFlags(UpsmonConfiguration::NotifyType type, unsigned short flags) {
+static std::string serializeNotifyFlags(UpsmonConfiguration::NotifyType type, unsigned int flags) {
 	static const NotifyFlagsStrings::FlagStrings::const_iterator ignore_str_iter =
 		NotifyFlagsStrings::flag_str.find(UpsmonConfiguration::NOTIFY_IGNORE);
